Move window related types into their own module
